refactored component names and deleted unused imports
